feat: Add cmdlet Write-ADOArtifact for uploading files to Azure DevOps artifacts
This commit adds a new cmdlet called Write-ADOArtifact to the AzureDevOpsLogging module. The cmdlet allows users to upload a local file into a file container folder in Azure DevOps artifacts and optionally publish it as an artifact. This functionality is useful for automating the process of uploading files as part of a CI/CD pipeline.

Signed-off-by: Marius Solbakken Mellum <marius.solbakken@fortytwo.io>